WIP Event CRUD - READ DETAILS
Implemented a primitive form for the EventCreate view. Letting Django do most of the heavy lifting here (thank you Keegan).
Modified urls.py to change the event list url to singular and add the create-event route. This change is also reflected in layout.html
Added the form into the event_create template.

Next step is to modify the EventList view to support the POST coming from the form.