Tentative styling for events Create and Update
Added HTML/CSS to style the create event and update event pages. I went for a centralized look for the form, but I'm not sure yet if this is the way to go. I also tweaked the back button on the Update page to re-route to the details page for the event.